requirements.txt
pyqt5~=5.15.10
pillow~=10.1.0
opencv-python~=4.8.0.74
matplotlib~=3.7.2
tensorflow~=2.13.0
numpy~=1.26.1
SQLAlchemy~=2.0.21
seaborn~=0.12.2
scikit-learn~=1.3.0